Move to the new traveler
